requirements.txt
Django>=4.2,<5.0
gunicorn>=21.2
psycopg[binary]>=3.1
python-dotenv>=1.0
